Incase we need full screen code is commented.

+++ AuthorRetriever/src/main/java/com/neu/msd/AuthorRetriever/constants/SceneContants.java
@@ -1,5 +1,9 @@
 package com.neu.msd.AuthorRetriever.constants;
 
+import javafx.geometry.Rectangle2D;
+import javafx.stage.Screen;
+import javafx.scene.Scene;
+
 public class SceneContants {
 	
 	public static final String SEARCH = "search";
@@ -11,6 +15,7 @@ public class SceneContants {
 	public static final String RESULT_TITLE = "RESULT";
 	public static final String AUTHOR_TITLE = "AUTHOR PROFILE";
 	
-	public static final int SCENE_LENGTH = 1000;
-	public static final int SCENE_WIDTH = 1000;
+	//public static final Rectangle2D screenBounds = Screen.getPrimary().getVisualBounds();	
+	public static final double SCENE_LENGTH = 1000;
+	public static final double SCENE_WIDTH = 1000;
 }
